+++ E-Commerce/src/main/java/com/uade/tpo/E_Commerce/repository/Seller_UserRepository.java
@@ -29,12 +29,12 @@ public interface Seller_UserRepository extends JpaRepository<Seller_User, Long> 
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO Seller_User(id_user, cuit, company_name, description, state,photo_url) VALUES (?1, ?2, ?3, ?4, ?5,?6)", nativeQuery = true)
-    int insertUser(long id_user, int cuit, String companyName, String description, String state,String photo_url);
+    int insertUser(long id_user, long cuit, String companyName, String description, String state,String photo_url);
 
     @Modifying
     @Transactional
     @Query(value = "UPDATE Seller_User SET cuit = ?2, company_name = ?3, description = ?4, state = ?5 WHERE id_user = ?1", nativeQuery = true)
-    int updateUser(long id, int cuit, String companyName, String description, String state);
+    int updateUser(long id, long cuit, String companyName, String description, String state);
 
     @Modifying
     @Transactional
